feat: notification_handler interface 정의
