+++ pennyway-infra/src/main/java/kr/co/pennyway/infra/common/event/NotificationEventHandler.java
@@ -1,40 +1,12 @@
 package kr.co.pennyway.infra.common.event;
 
-import com.google.api.core.ApiFuture;
-import kr.co.pennyway.infra.client.google.fcm.FcmManager;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.scheduling.annotation.Async;
-import org.springframework.stereotype.Component;
-import org.springframework.transaction.event.TransactionalEventListener;
-
-import java.util.concurrent.Executors;
-
 /**
- * FCM 푸시 알림을 처리하는 핸들러
+ * FCM 푸시 알림을 처리하는 핸들러 타입 인터페이스
+ * <p>
+ * 테스트 환경에서 Fcm Config 빈 생성을 막기 위해 사용하며, 테스트 시 해당 인터페이스를 구현한 Mock 객체를 사용한다.
+ *
+ * @author YANG JAESEO
+ * @since 2024-07-09
  */
-@Slf4j
-@Component
-@RequiredArgsConstructor
-public class NotificationEventHandler {
-    private final FcmManager fcmManager;
-
-    @Async
-    @TransactionalEventListener
-    public void handleEvent(NotificationEvent event) {
-        log.debug("handleEvent: {}", event);
-        ApiFuture<?> response = fcmManager.sendMessage(event);
-
-        if (response == null) {
-            return;
-        }
-
-        response.addListener(() -> {
-            try {
-                log.info("Successfully sent message: " + response.get());
-            } catch (Exception e) {
-                log.error("Failed to send message: " + e.getMessage());
-            }
-        }, Executors.newCachedThreadPool()); // FIXME: 알림이 매우 많은 경우 out of memory 발생 가능성 있음 (Thread pool size 제한 필요)
-    }
+public interface NotificationEventHandler {
 }
